separated stuffs into different folders and stuffs
