+++ src/components/Button.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import './button.css'
+import '../styles/button.css'
 import { Link } from 'react-router-dom'
 
 const STYLES = ['btn--primary', 'btn--outline']
