+++ src/components/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Route, Redirect } from 'react-router-dom'
-import { useAuth } from './contexts/AuthContext'
+import { useAuth } from '../contexts/AuthContext'
 
 export default function PrivateRoute({ component: Compnent, ...rest }) {
     const { currentUser } = useAuth()
